Move db, auth/jwks, spec, secrets, cli/backup, apierr into common/
Continues the config-service extraction (step 2 of 6). The common/
sub-module now contains everything the config-service repo will need
to import without pulling in identity's internal packages:

  common/db          — generic SQLite opener + migration runner
  common/apierr      — shared sentinel errors (ErrNotFound, ErrConflict, …)
  common/auth        — TokenClaims/ServiceTokenClaims types, ErrToken*,
                       JWKSVerifier, TokenParser interface
  common/spec        — YAML→JSON Converter for OpenAPI specs
  common/secrets     — DB-managed JWT signing key + session secret
  common/cli/backups — R2 backup list/restore CLI operations

Identity's internal packages become thin shims: type aliases and var
assignments pointing at the common types so no downstream callers need
import changes. errors.Is() keeps working because alias vars share the
same pointer as the canonical common sentinels.
