+++ go.mod
@@ -10,9 +10,6 @@ require (
 	github.com/sweeney/identity/common v0.0.0-00010101000000-000000000000
 	go.uber.org/mock v0.6.0
 	golang.org/x/crypto v0.49.0
-	golang.org/x/sync v0.20.0
-	gopkg.in/yaml.v3 v3.0.1
-	modernc.org/sqlite v1.46.2
 	rsc.io/qr v0.2.0
 )
 
@@ -52,9 +49,12 @@ require (
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/remyoudompheng/bigfft v0.0.0-20230129092748-24d4a6f8daec // indirect
 	github.com/x448/float16 v0.8.4 // indirect
+	golang.org/x/sync v0.20.0 // indirect
 	golang.org/x/sys v0.42.0 // indirect
 	golang.org/x/time v0.15.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 	modernc.org/libc v1.70.0 // indirect
 	modernc.org/mathutil v1.7.1 // indirect
 	modernc.org/memory v1.11.0 // indirect
+	modernc.org/sqlite v1.46.2 // indirect
 )
