+++ internal/spec/spec.go
@@ -2,10 +2,8 @@ package spec
 
 import (
 	_ "embed"
-	"encoding/json"
-	"sync"
 
-	"gopkg.in/yaml.v3"
+	commonspec "github.com/sweeney/identity/common/spec"
 )
 
 //go:embed openapi.yaml
@@ -15,38 +13,17 @@ var YAML []byte
 var ConfigYAML []byte
 
 var (
-	jsonOnce sync.Once
-	jsonData []byte
-	jsonErr  error
-
-	configJSONOnce sync.Once
-	configJSONData []byte
-	configJSONErr  error
+	identityConv *commonspec.Converter
+	configConv   *commonspec.Converter
 )
 
-// JSON returns the identity OpenAPI spec as JSON, converted from the
-// embedded YAML. The conversion is performed once and cached.
-func JSON() ([]byte, error) {
-	jsonOnce.Do(func() {
-		var v any
-		if err := yaml.Unmarshal(YAML, &v); err != nil {
-			jsonErr = err
-			return
-		}
-		jsonData, jsonErr = json.MarshalIndent(v, "", "  ")
-	})
-	return jsonData, jsonErr
+func init() {
+	identityConv = commonspec.NewConverter(YAML)
+	configConv = commonspec.NewConverter(ConfigYAML)
 }
 
+// JSON returns the identity OpenAPI spec as JSON, converted from the embedded YAML.
+func JSON() ([]byte, error) { return identityConv.JSON() }
+
 // ConfigJSON returns the config service OpenAPI spec as JSON.
-func ConfigJSON() ([]byte, error) {
-	configJSONOnce.Do(func() {
-		var v any
-		if err := yaml.Unmarshal(ConfigYAML, &v); err != nil {
-			configJSONErr = err
-			return
-		}
-		configJSONData, configJSONErr = json.MarshalIndent(v, "", "  ")
-	})
-	return configJSONData, configJSONErr
-}
+func ConfigJSON() ([]byte, error) { return configConv.JSON() }
